pyproject.toml
[project]
name = "dj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "django-guardian>=2.4.0",
    "django-ninja>=1.3.0",
    "django>=5.1.2",
    "elasticsearch>=8.15.1",
    "logfire[django]>=2.1.2",
    "pyjwt>=2.9.0",
]

[tool.uv]
dev-dependencies = [
    "django-stubs>=5.1.0",
    "mypy>=1.13.0",
    "ruff>=0.7.1",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "backend.settings"
